+++ include/yafth/core/input_handler_interface.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <yafth/core/types.h>
+
+namespace yafth::core
+{
+
+class input_handler_interface
+{
+    virtual input_type get_input() = 0;
+    virtual ~input_handler_interface() = default;
+};
+
+} // namespace yafth::core
